requirements.txt
openai==0.28.0
matplotlib==3.7.2
numpy==1.25.2
networkx==3.1
python-dotenv==1.0.0
